feat(dataplane): move existing code
Moves existing code outside of dataplane.rs in order to support
distinct packet reception drivers.
