+++ dataplane/src/drivers/mod.rs
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Open Network Fabric Authors
+
+#![allow(unused)]
+
+pub mod dpdk;
+pub mod kernel;
